+++ CMP104CyberPet/src/GUI/GUIScreen.cpp
@@ -34,12 +34,12 @@ void GUIScreen::OnArrowKey(GUIScreen::ArrowDirection dir)
 	case GUIScreen::ArrowDirection::Left:	
 	{
 		m_SelectedButton--;
-		if (m_SelectedButton == -1) m_SelectedButton = m_Buttons.size() - 1;
+		if (m_SelectedButton == -1) m_SelectedButton = (int)m_Buttons.size() - 1;
 	}; break;
 	case GUIScreen::ArrowDirection::Right:
 	{
 		m_SelectedButton++;
-		if (m_SelectedButton == m_Buttons.size()) m_SelectedButton = 0;
+		if (m_SelectedButton == (int)m_Buttons.size()) m_SelectedButton = 0;
 	}; break;
 	}
